requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.2
openai>=1.35.0
pyyaml>=6.0
